thermal: adc-tm: Fix uninitialized variable warning
In adc_tm5_thr_update, 'chip' is not initialized before it gets sent to
adc_tm5_get_btm_idx() but then it turns out that adc_tm5_get_btm_idx
doesn't even use the passed in pointer so nothing serious was happening
outside of the compiler warning.  Initialize the variable to fix the
warning but a wider fix to remove the variable is warranted in the
future.

 drivers/thermal/qcom/adc-tm5.c:265:
   'chip' is used uninitialized in this function [-Werror=uninitialized]

Fixes: bda25621 ("thermal: adc-tm: Add ADC_TM driver snapshot")
Change-Id: Ic0dedbad4eab1934bd6920beca2530bd870f5af4
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>