irq: wakeup logging for bad wakes
Multiple device families have hit issues recently in which recurring,
undesired kernel wakeups occur that only get logged as "unknown" by
BatteryStats.  Investigation has revealed that this can occur for a few
different reasons, including:
* driver configuring IRQ with IRQF_NO_SUSPEND when it is actually
  wakeup capable and should use enable_irq_wake() instead
* driver configuring IRQ with both IRQF_NO_SUSPEND and enable_irq_wake()
  even though the kernel documentation clearly states that that is not
  appropriate
* assertion of unmapped HW IRQs
* wakeups handled by TZ (e.g. watchdog, which leaves no evidenc1e of the
  wake reason once TZ handling has completed)

When these wakeups occur, the wakeup_reason logging in the kernel
doesn't capture the precipitating event, which results in the sysfs node
containing last wake reason remaining empty.  When BatteryStats sees
that, it logs the wakeup reason as "unknown". The only other two
formats that it supports (when a reason is present) are:
* "Abort: <reason>" - indicates that an attempt to suspend was aborted
  for the specified reason
* "<IRQ> <name>" - indicates that an interrupt with the specified
  virtual (_not_ HW) IRQ# and action name woke the kernel

This CL includes changes that address the second and third scenarios
listed above (i.e. _not_ the case where a driver uses IRQF_NO_SUSPEND
when it should use enable_irq_wake()).  It adds a new
log_bad_wake_reason() function to our existing wakeup_reason code, adds
a call to that function from the GIC driver when an unmapped HW IRQ is
detected (which should pretty much never occur on a healthy device),
and adds a call to irq_may_run() - which executes for virtually _every_
interrupt - in cases where a non-wakeup-armed interrupt turns out to
be misconfigured with IRQF_NO_SUSPEND and enable_irq_wake().

Bug: 127863679
Bug: 127862997
Test: Tested over a WE on a device
Change-Id: Ib4101fd74d6dae6b02266b7534cb2a6d1869080f
Signed-off-by: Kelly Rossmoyer <krossmo@google.com>
Signed-off-by: Thierry Strudel <tstrudel@google.com>
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>