mm: oom_kill: Kill the task causing the OOM
This commit is experimental. There are some risks involved.

The logic behind enabling this flag is that by killing the task that
caused the OOM condition rather than scanning the entire tasklist to use
heuristics to kill a task, the expensive tasklist scan is avoided.

The main opposing argument against using this strategy of killing is
that it's likely that the task that causes the OOM is an essential
foreground task, and killing it will interfere with the end-user. More
testing is required to determine if this is still an issue with LMKD on
newer Android systems, since OOM conditions are far less likely with an
efficient LMK.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>
Signed-off-by: UtsavisGreat <utsavbalar1231@gmail.com>