ion: remove cpu binding of pool refill kthreads
Remove the cpu affinity of ion pool refill kthreads which were
binded to min capability cpus, since there was no observable
power impact with when not affined.

Note: It was found that the code to create cpumask for min
capability cpus was broken and kthreads were free to run on
any cpus. No observable power impact was seen, hence better
off with removing this code entirely.

Change-Id: If6a32d9e4d3c576e63e3b0ec405be1fd4639a5e0
Signed-off-by: Sudarshan Rajagopalan <sudaraja@codeaurora.org>
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>