arm64: asm: spinlock: Add trailing NOP in spin_lock_trylock
Commit 00d37423ceaf from Qualcomm has removed the prfm instruction
in the spin_lock_trylock LL/SC code in order to solve a possible
hang in the event of a spin trylock during a timing change or an
interrupt trigger and also in order to increase the lock performance
on Cortex A53/A57 when the memory type is shareable.

While that commit actually solves a corner case issue and actually
enhances performance on these core designs, it's badly performed
because this does break the alternatives code in Linux which do
require the two INSNs to be of exactly the same length.

To solve this issue, add a NOP at the beginning of the trylock
LL/SC code: this makes sure that both the LSE and LL/SC flows
are of exactly the same size, hence solving "strange" assembler
issues like this one:
    spinlock.s:892: Error: attempt to move .org backwards

Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>