usb: gadget: f_qdss: Remove channel from list in free_inst
Currently the driver does not remove the channel corresponding to
an instance when freeing it. This leads to re-creation of a
channel failing because the older channel with the same name
still exists. Eventually, the function does not work.
Fix this by removing the channel from usb_qdss_ch_list as a part
of free_inst.

Change-Id: I114c0a0dcad5591ada50afbb28dd29dad6bb196f
Signed-off-by: Ajay Agarwal <ajaya@codeaurora.org>
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>